feat: add router for page routing
